imageactions: Use basename for creating thumbnail
Doing a split on . for the full path of the input file is not going to
work out well if one of the parent directories of the input file has a
dot in it.

So let's say if we have an input file with the following path:

/foo.bar/shiny_image.png

The basename of the resulting thumbnail would be:

foo.thumbnail_XXXxYYY.png

So instead of using the full path of the input file for determining the
first path of the new thumbnail file name we now solely use its
basename.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>