helpers: Use native Python to iterate $PATH
Using ls and tr means that we have two external commands just to iterate
through the $PATH environment variable.

This not only is kinda pointless to do (we *can* do this in Python
without forking off) but is also error-prone because it relies not only
on ls/tr being there but also on their outputs being correct.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>