main: Don't use sys.argv in test mode
If you run nosetests with arguments, the argument parser of vimiv picks
it up and execution may fail depending on which arguments are passed.

We now explicitly pass sys.argv to the main() function via
vimiv/vimiv.py and via main_test.py, so that if we call the main()
function via code we *never* provide arguments implicitly.

This also makes the arguments kvarg of parse_args() mandatory, because
there is no other code in the codebase that doesn't explicitly pass
arguments.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>