Rebase vimiv as a Gtk.Application
The main class is now a Gtk.Application instead of a Gtk.Window which makes much
more sense. The masses of attributes in the form of vimiv.thumbnail where
put into a dictionary widgets which can be accessed via self["widgetname"].

Tests were rewritten to follow the new codebase.

Here a new wrapper around TestCase called VimivTestCase was introduced which
generates the vimiv application in setUpClass and destroys it in tearDownClass.