Write a new completion model
Completions are now displayed in a GtkTreeView above the entry instead of in a
Gtk.Label. This allows the usage of native Gtk filtering instead of writing the
completion filter by hand. Additionally information on internal commands can be
displayed next to the command itself.

To make the information work well the manpage on completions was slightly
modified. Mainly sentences were shortened and some missing commands were added.

To read the info from the manpage a new function in helpers.py was introduced.

Downside:
    Completion of external commands was dropped because it is much to slow with
    native Gtk filtering.