Deprecate combined keybinding sections like IM_LIB
It is difficult to get an overview of all keybindings in one mode with
the different sections that apply to it. Therefore sort the default
keybinding alphabetically according to command names and remove all
combined sections.

They will still work with user configurations but a deprecation warning
is printed to stdout on startup.