Load images and animations async with ImageLoader
The actual loading of images was moved to an extra class called
ImageLoader. This class deals with the actual reading process in an
extra thread adding a call to Image.update() as soon as the image is
finished. This does not change much for normal images, except for that
the UI stays responsive. Animated Gifs on the other hand are now
displayed as soon as possible removing annoying waiting times for large
Gifs (with many frames).