Make use of the newly implemented SettingStorage
This commit became rather large :(

Lots of class attributes were replaced by the value stored in the
SettingStorage. This allows access from everywhere and, much more
important, the use of a global :set command.

The :set command can change values in the SettingStorage. If values are
changed, the storage emits a "changed" signal to which the base classes
can connect and react appropriately. For example :set show_hidden!
toggles the "show_hidden" setting in storage. The library then reacts
via reload.

This lead to a rename of quite a few settings and commands.

Still TODO:
- Get aliases running again